chore: 프로젝트 구조 변경
- domain - entity, repository
- interfaces - controller, dto
