--- charlee-web/src/main/java/com/efa73/charleeweb/user/domain/Role.java
@@ -1,5 +1,0 @@
-package com.efa73.charleeweb.user.domain;
-
-public enum Role {
-    ADMIN, COMPANY, USER
-}
