+++ charlee-web/src/main/java/com/efa73/charleeweb/user/domain/entity/Role.java
@@ -1,0 +1,5 @@
+package com.efa73.charleeweb.user.domain.entity;
+
+public enum Role {
+    ADMIN, COMPANY, USER
+}
